factor out spartan-based snark into a separate module
